building new build base image
