+++ autorest/Dockerfile
@@ -1,32 +1,3 @@
-FROM iusdev.azurecr.io/web-build:latest AS build
-
-RUN apt-get update
-RUN apt-get install -y \
-	git \
-	wget \
-	curl \
-	build-essential \
-	zlib1g-dev \
-	libncurses5-dev \
-	libgdbm-dev \
-	libnss3-dev \
-	libssl-dev \
-	libreadline-dev \
-	libffi-dev
-
-# Set Python version
-ARG PYTHON_VERSION='3.7.0'
-# Set pyenv home
-ARG PYENV_HOME=/root/.pyenv
-
-# Install pyenv, then install python versions
-RUN git clone --depth 1 https://github.com/pyenv/pyenv.git $PYENV_HOME && \
-    rm -rfv $PYENV_HOME/.git
-
-ENV PATH $PYENV_HOME/shims:$PYENV_HOME/bin:$PATH
-
-RUN pyenv install $PYTHON_VERSION
-RUN pyenv global $PYTHON_VERSION
-RUN pip install --upgrade pip && pyenv rehash
+FROM iusdev.azurecr.io/ius-build-base:latest AS build
 
 RUN npm install -g autorest
