--- buildtools/version-checker/Dockerfile
@@ -1,10 +1,0 @@
-FROM python:3.8
-
-WORKDIR /data
-
-COPY ./bannedversions.json /data/
-COPY ./versionchecker.py /data/
-COPY ./requirements.txt /data
-
-RUN pip install -r requirements.txt
-
