--- buildtools/version-checker/bannedversions.json
@@ -1,10 +1,0 @@
-{
-    "IdentityModel" : {
-        "MustBe" : "LessThan",
-        "Version" : "4.0.0"
-    },
-    "IdentityServer4" : {
-        "MustBe" : "LessThan",
-        "Version" : "3.0.0"
-    }
-}
