--- buildtools/version-checker/requirements.txt
@@ -1,1 +1,0 @@
-packaging==19.2
